requirements.txt
opencv-python>=4.12.0.88
streamlit>=1.47.1
tensorflow>=2.19.0
pillow
numpy
